Added mapPicker
