+++ client/app/components/group/store/_storeEditCreateForm/storeEditCreateForm.controller.js
@@ -1,17 +1,9 @@
 class StoreEditCreateFormController {
-  constructor(Geocoding, $stateParams, CurrentGroup, $scope) {
+  constructor($stateParams, $scope) {
     "ngInject";
     Object.assign(this, {
-      CurrentGroup,
-      Geocoding,
       $stateParams,
-      $scope,
-      mapCenter: {},
-      mapDefaults: {
-        scrollWheelZoom: false,
-        zoomControl: true,
-        dragging: true
-      }
+      $scope
     });
   }
 
@@ -19,26 +11,11 @@ class StoreEditCreateFormController {
     if (angular.isUndefined(this.data)) {
       Object.assign(this, {
         isCreate: true,
-        mapCenter: {
-          lat: this.CurrentGroup.value.latitude,
-          lng: this.CurrentGroup.value.longitude,
-          zoom: 12
-        },
         data: {
           group: this.$stateParams.groupId
         }
       });
-    } else {
-      this.trySetLocation(this.data);
     }
-
-    this.$scope.$on("leafletDirectiveMap.click", (event, e) => {
-      let item = {
-        latitude: e.leafletEvent.latlng.lat,
-        longitude: e.leafletEvent.latlng.lng
-      };
-      this.setMarker(item);
-    });
   }
 
   submit() {
@@ -58,43 +35,6 @@ class StoreEditCreateFormController {
       });
     });
   }
-
-  geoLookup() {
-    return this.Geocoding.lookupAddress(this.query);
-  }
-
-  setMarker(item) {
-    if (!this.marker || !this.marker.p) this.marker = { p: {} };
-    angular.copy({
-      lat: item.latitude,
-      lng: item.longitude,
-      message: item.address,
-      draggable: true
-    }, this.marker.p);
-  }
-
-  trySetLocation(item) {
-    if (!item || !item.address ) return;
-    this.setMarker(item);
-    this.query = item.address;
-    this.mapCenter.zoom = 15;
-    this.mapCenter.lat = item.latitude;
-    this.mapCenter.lng = item.longitude;
-    Object.assign(this.data, item);
-  }
-
-  updateOrDeleteIfEmpty(text) {
-    this.data.address = text;
-    if (!text) {
-      Object.assign(this.data, {
-        latitude: null,
-        longitude: null,
-        address: null
-      });
-    }
-  }
-
-
 }
 
 export default StoreEditCreateFormController;
