+++ client/app/components/group/store/_storeEditCreateForm/storeEditCreateForm.js
@@ -1,21 +1,15 @@
 import angular from "angular";
 import uiRouter from "@uirouter/angularjs";
 import storeEditCreateFormComponent from "./storeEditCreateForm.component";
-import Geocoding from "services/geocoding/geocoding";
 import StoreService from "services/store/store";
 import markdownInput from "components/_markdownInput/markdownInput";
-import Group from "services/group/group";
-import "leaflet";
-import "ui-leaflet";
-import "leaflet/dist/leaflet.css"; // looks in node_modules
+import mapPicker from "components/_mapPicker/mapPicker";
 
 let storeEditCreateFormModule = angular.module("storeEditCreateForm", [
   uiRouter,
-  Geocoding,
   StoreService,
   markdownInput,
-  Group,
-  "ui-leaflet"
+  mapPicker
 ])
 
 .component("storeEditCreateForm", storeEditCreateFormComponent)
